Renable Roundstart Pilgrim + Adds Outlander Virtue
Allows for Pilgrim Roundstart.

Additionally, adds the Outlander virtue. This virtue marks you as a foreigner, gives you a small pouch of rations, and gives you 2 levels in traversal skills from your long trek to The Ridge. These skills cannot exceed 4, and so are of little value to most combat roles. Command Roles are incompatible with the Outlander Virtue.