Fuck it; Some Original Stuff & Fixes
Resprites The Progress Bar back to /tg/ standard; but in the style of parchment! I thought it was neat; you can't blame me too heavily.

Resprites random spawners to /tg/ standard as well. They're no longer purple squares.
Why is this bundled in the port PR? I wanted to port the AI improvements due to an uptick in pathfinding abuse atm; but that's not ready yet and I didn't want to wait on critical logging.

So. Yanno. Fuck it; right? Nobody's gonna tell me to atomize my own PR when I'm gonna merge it myself. >:)